% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate_camera_orbit}
\alias{animate_camera_orbit}
\title{Attach orbit animation metadata to a scene}
\usage{
animate_camera_orbit(scene, frames = 36L, axis = "z", step = 5)
}
\arguments{
\item{scene}{An \code{xgeo_scene} object.}

\item{frames}{Number of animation frames.}

\item{axis}{Rotation axis. One of \code{"x"}, \code{"y"}, or \code{"z"}.}

\item{step}{Degrees per frame.}
}
\value{
The updated \code{xgeo_scene}.
}
\description{
Attach orbit animation metadata to a scene
}
