% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgeo_data_coerce.R
\name{as_xgeo_data}
\alias{as_xgeo_data}
\title{Coerce inputs to \code{xgeo_data}}
\usage{
as_xgeo_data(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Passed to method-specific implementations.}
}
\value{
A \code{xgeo_data} object.
}
\description{
Coerce inputs to \code{xgeo_data}
}
