% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_surface.R
\name{geom_xgeo_surface}
\alias{geom_xgeo_surface}
\title{Add a generic surface layer}
\usage{
geom_xgeo_surface(
  scaling = 1,
  alpha = 0.7,
  smooth = FALSE,
  low_fill = "#20639B",
  high_fill = "#D1495B"
)
}
\arguments{
\item{scaling}{Vertical scale multiplier for explanation values.}

\item{alpha}{Surface alpha.}

\item{smooth}{Whether to apply a simple neighborhood average before
rendering.}

\item{low_fill}{Color used near the lower value range.}

\item{high_fill}{Color used near the upper value range.}
}
\value{
An \code{xgeo_layer} object.
}
\description{
Add a generic surface layer
}
