% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_scene.R
\name{read_xgeo_scene}
\alias{read_xgeo_scene}
\title{Read a scene from JSON}
\usage{
read_xgeo_scene(path)
}
\arguments{
\item{path}{Path to a JSON scene file.}
}
\value{
An \code{xgeo_scene} object.
}
\description{
Read a scene from JSON
}
