% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_webgl.R
\name{render_xgeo_layer}
\alias{render_xgeo_layer}
\title{Render one \code{xgeo_layer}}
\usage{
render_xgeo_layer(data, layer, scene, ...)
}
\arguments{
\item{data}{Long-table data from an \code{xgeo_data} object.}

\item{layer}{Layer object to render.}

\item{scene}{Parent scene.}

\item{...}{Reserved for future extension methods.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Render one \code{xgeo_layer}
}
