% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgeo_state_class.R
\name{validate_xgeo_state}
\alias{validate_xgeo_state}
\title{Validate an \code{xgeo_state} object}
\usage{
validate_xgeo_state(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
\code{x}, invisibly, when validation succeeds.
}
\description{
Validate an \code{xgeo_state} object
}
\examples{
state <- xgeo_state(matrix(c(1, -1, 2, 0), nrow = 2))

validate_xgeo_state(state)
}
