% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgeo_state_accessors.R
\name{xgeo_attributes}
\alias{xgeo_attributes}
\title{Get attributes from an \code{xgeo_state}}
\usage{
xgeo_attributes(state)
}
\arguments{
\item{state}{An \code{xgeo_state} object.}
}
\value{
The \code{attributes} field.
}
\description{
Get attributes from an \code{xgeo_state}
}
\examples{
state <- xgeo_state(matrix(c(1, 2, 3, 4), nrow = 2))

names(xgeo_attributes(state))
}
