% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgeo_state_accessors.R
\name{xgeo_geometry}
\alias{xgeo_geometry}
\title{Get geometry from an \code{xgeo_state}}
\usage{
xgeo_geometry(state)
}
\arguments{
\item{state}{An \code{xgeo_state} object.}
}
\value{
The \code{geometry} field.
}
\description{
Get geometry from an \code{xgeo_state}
}
\examples{
state <- xgeo_state(matrix(c(1, 2, 3, 4), nrow = 2))

xgeo_geometry(state)$points
}
