% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgeo_state_accessors.R
\name{xgeo_selection}
\alias{xgeo_selection}
\title{Get selection from an \code{xgeo_state}}
\usage{
xgeo_selection(state)
}
\arguments{
\item{state}{An \code{xgeo_state} object.}
}
\value{
The \code{selection} field.
}
\description{
Get selection from an \code{xgeo_state}
}
