% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgeo_state_coerce.R
\name{as_xgeo_state}
\alias{as_xgeo_state}
\title{Coerce inputs to \code{xgeo_state}}
\usage{
as_xgeo_state(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Passed to method-specific implementations.}
}
\value{
An \code{xgeo_state} object.
}
\description{
Coerce inputs to \code{xgeo_state}
}
