% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_xgeo_state.R
\name{read_xgeo_state}
\alias{read_xgeo_state}
\title{Read a backend state from JSON}
\usage{
read_xgeo_state(path)
}
\arguments{
\item{path}{Path to a JSON state file.}
}
\value{
An \code{xgeo_state} object.
}
\description{
Read a backend state from JSON
}
