% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgeo_data_class.R
\name{validate_xgeo_data}
\alias{validate_xgeo_data}
\title{Validate a \code{xgeo_data} object}
\usage{
validate_xgeo_data(x)
}
\arguments{
\item{x}{An object to validate.}
}
\value{
\code{x}, invisibly, when validation succeeds.
}
\description{
Validate a \code{xgeo_data} object
}
