% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgeo_state_accessors.R
\name{xgeo_indices}
\alias{xgeo_indices}
\title{Get indices from an \code{xgeo_state}}
\usage{
xgeo_indices(state)
}
\arguments{
\item{state}{An \code{xgeo_state} object.}
}
\value{
The \code{indices} field.
}
\description{
Get indices from an \code{xgeo_state}
}
