% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgeo_state_accessors.R
\name{xgeo_metadata}
\alias{xgeo_metadata}
\title{Get metadata from an \code{xgeo_state}}
\usage{
xgeo_metadata(state)
}
\arguments{
\item{state}{An \code{xgeo_state} object.}
}
\value{
The \code{metadata} field.
}
\description{
Get metadata from an \code{xgeo_state}
}
